Increase Map size dimension units from byte to ushort

+++ src/Models/IMap.cs
@@ -10,17 +10,17 @@ namespace LemballEditor.Models
         /// <summary>
         /// Number of X tiles
         /// </summary>
-        byte XTiles { get; }
+        ushort XTiles { get; }
 
         /// <summary>
         /// Number of Y tiles
         /// </summary>
-        byte YTiles { get; }
+        ushort YTiles { get; }
 
         /// <summary>
         /// The number of tiles in the level
         /// </summary>
-        ushort TileCount { get; }
+        int TileCount { get; }
 
         /// <summary>
         /// Returns an iterator that iterates through the tiles sequentially
@@ -34,7 +34,7 @@ namespace LemballEditor.Models
         /// <param name="xTile">0-based xTile coordinate</param>
         /// <param name="yTile">0-based yTile cordinate</param>
         /// <param name="tile">The tile to place at the given position</param>
-        void SetTile(byte xTile, byte yTile, Tile tile);
+        void SetTile(ushort xTile, ushort yTile, Tile tile);
 
         /// <summary>
         /// Get the tile at the given coordinate, or null if none has been set there
@@ -42,6 +42,6 @@ namespace LemballEditor.Models
         /// <param name="xTile">0-based xTile coordinate</param>
         /// <param name="yTile">0-based yTile cordinate</param>
         /// <returns>Tile, or null if none set</returns>
-        Tile GetTile(byte xTile, byte yTile);
+        Tile GetTile(ushort xTile, ushort yTile);
     }
 }
